Assignments/assignmentFour/requirements.txt
flask
mysql-connector-python
tabulate
requests
json